feat: confirmation modal before deleting all lists

+++ src/components/Buttons.js
@@ -1,20 +1,29 @@
 import { Button } from "@mui/material";
 import AddIcon from "@mui/icons-material/Add";
 import DeleteIcon from "@mui/icons-material/Delete";
-import React from "react";
-import { todoActions } from "../store/index.js"
-import { useDispatch } from "react-redux"
+import React, { useState } from "react";
+import { todoActions } from "../store/index.js";
+import { useDispatch } from "react-redux";
+import Alert from "../ui/Alert.js";
 
 const Buttons = () => {
-    const dispatch = useDispatch();
+  const dispatch = useDispatch();
+  const [alertOpen, setAlertOpen] = useState(false);
 
-    const addNewListHandler = () => {
-        dispatch(todoActions.addTodoList());
-    }
+  const addNewListHandler = () => {
+    dispatch(todoActions.addTodoList());
+  };
 
-    const deleteAllListsHandler = () => {
-        dispatch(todoActions.deleteAllTodoLists());
+  const deleteAllListsHandler = () => {
+    setAlertOpen(true);
+  };
+
+  const closeHandler = (close) => {
+    if (close) {
+      dispatch(todoActions.deleteAllTodoLists());
     }
+    setAlertOpen(false);
+  }
 
   return (
     <div>
@@ -42,6 +51,7 @@ const Buttons = () => {
       >
         Delete All Lists
       </Button>
+      <Alert open={alertOpen} onClose={closeHandler} />
     </div>
   );
 };
